requirements.txt
amqp==5.1.0
asgiref==3.5.0
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
cfgv==3.3.1
charset-normalizer==2.0.12
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Deprecated==1.2.13
distlib==0.3.4
Django==3.2.12
django-celery-beat==2.2.1
django-celery-results==2.3.0
django-rest-framework==0.1.0
django-timezone-field==4.2.3
djangorestframework==3.13.1
filelock==3.6.0
identify==2.4.11
idna==3.3
kombu==5.2.4
mailersend==0.2.1
nodeenv==1.6.0
packaging==21.3
platformdirs==2.5.1
pre-commit==2.17.0
prompt-toolkit==3.0.28
pyparsing==3.0.7
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
redis==4.1.4
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
urllib3==1.26.8
vine==5.0.0
virtualenv==20.13.3
wcwidth==0.2.5
wrapt==1.14.0
